Support the union syntax in type aliases
Summary: We did not support the union syntax `A | B` in type aliases. This was because we implemented this as a pre-processing step that transforms with the union syntax `A | B` into the equivalent `Union[A, B]`, but only in places where a type annotation is expected. This diff adds that transformation when resolving type aliases.

Reviewed By: grievejia

Differential Revision: D35412442

fbshipit-source-id: de6d28899a47620c58268c2c5a8705381ca163d1