Store the intervals of receiver objects and if the call is on self for every call site in a model
Summary:
The post processing in SAPP requires the intervals in the models to have a particular meaning (so that it can avoid expanding traces that do not exist due to the restriction imposed by class intervals T105515289, D32336953) -- every call site in a model needs to be associated with the class interval of the receiver object's type. Since this is different from the meaning of intervals currently associated with a call site, this diff *additionally* stores the intervals of receiver objects.

Additionally, SAPP requires a boolean flag associated with every call site that indicates whether the call is on self.

Reviewed By: arthaud

Differential Revision: D35261673

fbshipit-source-id: 8c90528a5d1c27d92458fe4dc3e54b089527b372