config in ModuleTracker [3c] Use the configuration from ModuleTracker
Summary:
This commit updates the core logic for `update_this_and_all_preceding_environments`
to ignore the configuration passed in (except for the purposes of passing it down)
and instead use the configuration coming from `ModuleTracker`.

The function argument is still there, which is just to keep the commit small and
focused; I'll remove it in the next commit which will involve a lot of tiny updates
to tests.

Reviewed By: grievejia

Differential Revision: D35413226

fbshipit-source-id: 9de1390806a28620835fcec80e09ab9a84f9ccb8