config in ModuleTracker [2] pass it up to AstEnvironment
Summary:
This commit is at the heart of why we're moving `configuration` into
`ModuleTracker.t` - there are several functions in `AstEnvironment` that
we want to migrate, and it was distracting to be passing around
`configuration` unnecessarily.

Instead, `AstEnvironment` public functions should never directly
accept a configuration, we should always just get it from the module tracker.
As of this commit, no public function requires a configuration as input.

Reviewed By: grievejia

Differential Revision: D35413224

fbshipit-source-id: 0a6790f5507755e3a6af01887139297a169075e2