config in ModuleTracker [3b] Rearrange Environment.ml code
Summary:
In order to actually make use of the configuration from ModuleTracker
in `Environment.update_this_and_all_preceding_environments`, we need
to reorder the code.

I decided this is a good time to pick a good code order and also fix
some strange things:
- the function to construct a read_only environment from an upstream
  UpdateResult shouldn't be named `read_only`, that collides with
  the actual `read_only` function which the compiler can handle but
  typical humans can't (I got pretty confused!) so I renamed it
- All of the basic operations on EnvironmentTable.S.t can and should
  be defined at the top, before the `ReadOnly` submodule
- The two update functions ought to be next to one another

This is a pure rearrangement + local-only rename, no logic changes at
all in this commit.

Reviewed By: grievejia

Differential Revision: D35413227

fbshipit-source-id: 28751d420e3b8929bb69a7be58e5453af786cae2