config in ModuleTracker [3d] remove from update_this_and_all_preceding_environments
Summary:
This commit finishes off the switch from passing around `configuration` explicitly
to fetching it from `ModuleTracker` but changing the interface of environment updates
to no longer require a `configuration`.

Since the configuration was already unused it's a simple diff just stripping out
the argument; it touches a lot of code because of how many tests call this update
function.

Reviewed By: grievejia

Differential Revision: D35413225

fbshipit-source-id: a4186eb4a4a442fb72f3ad1c219b7f78a50be79c