extend ModuleTracker [0]: create the layout
Summary:
In order to facilitate clean handling of the filesystem (which has to be
abstracted) in order to power pyre-for-unsaved-editor-state, we want
to get all of the filesystem handling in one place.

One problem, though, is that the overlay data isn't really something we would
want to save; we need a name for the thing that we serialize when "saving"
a module tracker.

This diff renames the existing tracker data - which will soon be only a
part of what the tracker contains - to the "layout", which is a mostly-internal
concept: it's the pair of maps that we compute by scanning the filesystem
that represents the layout of modules.

The only place where the idea of a "layout" leaks is in the names of
the serialization methods; other things we want to store in the module
tracker (such as the function to get raw code and the configuration) are not
part of the layout and are not serialized.

Because the tracker is currently nothing but a layout this commit looks a little
awkward, but in the next I will add `configuration` to the tracker at
which point it will make more sense why we want to separate the "computed
layout information" from the "other stuff" in the tracker.

Reviewed By: grievejia

Differential Revision: D35402304

fbshipit-source-id: 4a193ca64d332ef204f6efe974cfa4d17f617e03