config in ModuleTracker [1] push configuration into ModuleTracker.t
Summary:
This commit extends ModuleTracker by storing configuration
inside of it.

After some discussion, we realized this makes sense - it
simplifies the interfaces of several operations if we can
stop passing configuration down, and moreover it more accurately
reflects the semantic reality of our environment stack, which
is created under the assumption of a single immutable
configuration.

We wanted to make this change before pushing file-reading
down because it allows us to have cleaner interfaces when we
make that upcoming change.

Reviewed By: grievejia

Differential Revision: D35402303

fbshipit-source-id: 072356ed5beca1b632128c94cfe3265f45ebd181