config in ModuleTracker [4] ModuleTracker.Layouts
Summary:
In review of the commit that created `ModuleTracker.layouts`,
it was suggested that we create a `Layouts` module and make a
`Layouts.t` type rather than using a `layouts` type.

That was a hard change to restack, but it's easy to do as
a cleanup step. This diff has no logic changes, but moves all
layout-specific logic (create and update) to live in a `Layouts`
module. The resulting ModuleTracker codebase is significantly
better modularized now since this forced more aggressive grouping
of related functions.

Reviewed By: grievejia

Differential Revision: D35420291

fbshipit-source-id: 4749b0b44dfa3b12799dd6eb9f128cc46afcaf49