Remove --incremental-style
Summary:
It already seems to be the case that this field - which has been
marked as deprecated for at least a year and a half (it was already
deprecated in October 2020) - is not used.

Let's get rid of it unless someone sees a reason not to.

**Context on why I'm looking at this in the first place**

The context is that I'm trying to understand how configuration
actually threads through environment building. The main way it
does so is via the `incremental_style` config field, but I'm
currently under the impression that we could remove that field
from configuration entirely, putting it in the updates instead, and
set it deterministically via the control flow rather than by hacking
up the configuration.

It's not clear whether that's worth doing, but it is pretty clear that
it's worth making more obvious that this isn't actually a command-line
controllable option anymore.

Reviewed By: grievejia

Differential Revision: D35360560

fbshipit-source-id: c3e05d4197d9a25552b1c31dd78d96e3cf9a3eab