Add integration tests showing a bug in the order of sanitizers and named transforms
Summary:
# Context
In this diff stack, we implement proper interaction between sanitizers and taint transforms.
For short, we want taint transforms to invalidate all previous sanitizers, as if transforming was creating a new taint.

For instance:
```
def foo():
  x = source_a()
  y = sanitize_x(x)
  z = transform_t(y)
  sink_x(z)
```
This should be a valid issue even though `sanitize_x` should sanitize for the sink `X`, but the taint was transformed into `T`.

# This diff
First, we need to preserve the order of applied sanitizers in respect to taint transforms. This diff demonstrates that we do not preserve it properly when inferring taint-in-taint-out.

Reviewed By: alexkassil

Differential Revision: D35454636

fbshipit-source-id: fd9acfbfd416abc47f25f6edffb67219e5ed04bd