Fix a bug in the order of sanitizers and named transforms
Summary: During the forward and backward analysis, when handling a function call, we split transforms into sanitizers and named transforms and apply them in that order. This is not sound since sanitizers can interact with named transforms. They need to be applied at the same time. This is what this diff implements.

Reviewed By: alexkassil

Differential Revision: D35484441

fbshipit-source-id: c0b5303e55689cfe8030add743934d813ccb200e