Improve the handling on untracked type expression in call graph building [1/n]
Summary:
# Context
Most of pyre APIs can raise the exception `Analysis.ClassHierarchy.Untracked`, which happens when it finds a type that is not present in the class hierarchy.

# Problem
Right now, we catch this at the top level of call graph building, which means we will give up on analyzing a function if this exception is thrown at any time during call graph building. This can lead to false negatives.

# Solution
This is more tedious, but we should catch that exception at each call site and default to a sane behavior. It is also helpful to log information about what caused the error so it can be fixed.

This diff handles a specific call site that is causing issues with typeshed.

Reviewed By: alexkassil

Differential Revision: D35517689

fbshipit-source-id: d6a17b125a63bb89867f9151c1bea3be4b6354e1