flatten Memory [2] push from_string down
Summary:
This is the second commit of a series where I work on getting rid
of `Memory.{NoCache,WithCache}` and instead just using `SharedMemory`.
To do this, I need to either remove or push down each of the bits of logic
that currently live only in `Memory`.

This commit pushes down the `from_string` functionality; it turns out that
we already have this for every single key type in our program, so there's
no harm in just requiring it for the `SharedMemory` key module.

Reviewed By: grievejia

Differential Revision: D35502592

fbshipit-source-id: 2f1b5c804489e3ad6a450060d1a2085241a44b22