flatten Memory [3] Remove unmarshall function
Summary:
This commit is part of a series where I am working to get rid
of the current two-layer logic where global shared memory tables
are defined mostly in `SharedMemory` but have some logic in
`Memory`. Flattening the structure will make our code simpler to
navigate, and more importantly it will make it quite a bit easier
to implement overlayable shared memory because we won't have to
juggle as many different module types and functors.

As a result, I'm either eliminating aspects of the `Memory`
interface that aren't needed, or pushing them down into `SharedMemory`.

This commit eliminates the `unmarshall` function, which was actually
completely unused - it turns out all of the serialization is being
done by `SharedMemory`, which doesn't even have access to this function!
As a result, it's a no-op to strip it from our entire codebase.

Reviewed By: grievejia

Differential Revision: D35502593

fbshipit-source-id: 3d3ea5659b2aab83c0e4d6e57b15a8c8890c34f0