flatten Memory [1] Remove Memory.Key.out
Summary:
In order to support unsaved changes in Pyre, we need to make
shared memory overlayable. Doing this is significantly harder than
it needs to be right now because of the number of modules involved -
we mostly work with Memory.{WithCache,NoCache} but the underlying
interfaces that need to be replicated live in SharedMemory.

Flattening out the logic so that there's only one layer of modules for the
existing non-overlayable shared memory implementation will make things
much easier, so I'm working toward klling `Memory.{WithCache,NoCache}`
by taking the few bits of extra logic it has and either removing them if
unnecessary or pushing them down to `SharedMemory`.

This first commit removes `Key.out` - it turns out that every single key we
have is easy to convert back and forth with `string`, so there's no need to
allow the `Key.out` type to differ from `Key.t` - instead `from_string` can just
have type `string -> t`. Most of our modules work as-is for this, the rest of them
I've switched to use `Sexp` for the conversion.

Reviewed By: grievejia

Differential Revision: D35483680

fbshipit-source-id: 66dbbd2beb50ce94f4f79d31811445cf241b8b89