Print the exception backtrace
Summary:
We currently have a catch-all at the top level, but it doesn't print the backtrace of the exception because that trace is mostly lost due to using the `Lwt` asynchronous framework.

To ease debugging, let's print the exception and its backtrace within `StaticAnalysis`, so we get a full backtrace.

Reviewed By: stroxler

Differential Revision: D35562279

fbshipit-source-id: 68bf7e6e47546e25eab7328e4eec27c9c182c86f