Propagate exceptions from workers to the main process
Summary:
# Context
When an exception happens in a worker, it currently results into a `Worker_exited_abnormally` exception raised in the main process.
This makes it pretty hard to debug what actually happened.

# Solution
Let's propagate the exception from workers to the main process, as a string and a backtrace. This can be done by marshal-ing a result type between workers and the main process.

Reviewed By: stroxler

Differential Revision: D35567257

fbshipit-source-id: 37ba406507e6d0a5cc0a7debb5a9f552e1a04321