Use clearer name for submodules in AstEnvironment
Summary:
The name `updated_submodules` was not clearly indicating the semantics
here - module tracker will only include an implicit package module qualifier in
an update result if it is either removed (in which case it joins explicit
deletions in `removed_submodules`) or if it is new; no other kind of update
is possible for implicit modules. If a formerly implicit module becomes explicit
then it will be a `NewExplicit` and join `reparse_source_paths`.

Naming the implicit-module variable `new_submodules` makes this clearer.

NOTE eventually I'd like to stop calling these submodules and refer to them
instead as implcit modules; they aren't really submodules at all, they are
actually *super* modules since they are implied by the existence of modules
nested underneath them. So the name is pretty confusing! But that's less urgent
to me than making sure the update semantics are clear.

Reviewed By: grievejia

Differential Revision: D35538864

fbshipit-source-id: 8190acd05dc42f1333b73b598cd96fd97d2a9d97