Improve names in KeyTracker
Summary:
It was pretty weird to have a function called
`get_keys` that took `keys` as input, and conveyed absolutely
nothing about what was actually going on.

I think it might be better to rename these functions to
`get_*_names`, but that's debateable since they are in fact
keys of the dependency tracked tables.

What's not debateable is naming the input `qualifiers` helps
quite a bit - it clarifies that these are module names. Moreover,
`get_keys` was not a good name for the function that got
class names - `get_class_keys` makes it much clearer what we're
actually going to get back.

Reviewed By: grievejia

Differential Revision: D35561480

fbshipit-source-id: 764fdf0059bb4d22aed44da413071220568cbfff