Make control flow clearer in ModuleTracker update
Summary:
I had some trouble understanding what was happining in ModuleTracker
update because, in a local function that wasn't really necessary,
we were renaming `explicit_updates` and `implicit_updates` to
`explicits` and `implicits` and then shadowing the names with
a new type.

It's not that the code was so confusing, but removing the local
functions cuts down on the number of named things and makes it clearer
what's happening. Moreover, moving the logging statements to the top
also makes it clearer that we're logging the internal-typed updates
rather than the external-typed updates.

Reviewed By: grievejia

Differential Revision: D35538865

fbshipit-source-id: da882fb2f2c038f7bbb181fb26742258599cce2b