Lazy parsing: Push set_module_data into WriteOnly
Summary:
In order to have a good open-source experience, we want to make AstEnvironment
lazy. This may also eventually help with perf internally, because
- projects with deep dependency graphs where most of the dependencies aren't
  directly imported or exposed via class inheritance could skip a lot of work.
- we could potentially start initializing the LSP to only check open files,
  deferring the more expensive full check until later. This could be a big UX
  win in that we'd get code navigation in in-buffer errors *much* faster,
  although it's hard to say if we would do it, since then the problem becomes
  when to actually run a full check.
- we might get better parallelism if we don't have to parse everything ahead of
  time (it depends - we might get more duplicated work)

It turns out that it's not actually *that* hard to make AstEnvironment lazy;
mostly the problem is that it doesn't follow the same patterns as the rest of
the code so it takes longer to build context.

But as far as I can tell, all we really need is to:
- create a lazy loader in the unannotated global environment
  so that whenever we try to get some information
  inside a module (e.g. a class summary, a function definition, the list of
  unannotated globals, the list of all functions) we will populate all the
  information we need.
- modify AstEnvironment so that we only populate RawSources either when
  get_raw_source is called or when ModuleTracker informs us of an update to
  an already-parsed source.
- accept that get_all_xyz will, for a lazy-built environment, return
  only what has actually been loaded thus far (which will depend on what kind
  of end operations you do; for typecheck, it will wind up loading all
  direct imports plus any modules needed to resolve class ancestors, type
  aliases, and other recursively-resolved type information).

All of this can be made cleaner by following the Environment API all the
way down, which would make it simpler since then ModuleTracker could track
dependencies for a lazy AstEnvironment. I haven't yet decided whether to
do that refactor first or try to make the existing code lazy without clean
interfaces.

This commit works toward having a lazy loader by rolling together all of the
work needed to populate the unannotated global environment.

Reviewed By: grievejia

Differential Revision: D35561479

fbshipit-source-id: af1b4d83b27358b91c6a4b75ea1ff614187761e9