Refactor unannotatedGlobalEnvironmentTest.ml
Summary:
A flat-out majority of all test cases were in the mega-test-function
`test_update`. This is not good for doing a significant refactor, especially
because `assert_update` is very complicated, which means the tracebacks
and error messages tend to be harder to navigate than a typical assertion
would be.

So, in this commit I do the following:
- reorder the tests so that the order of definition matches the order
  in the test block at the bottom
- extract `assert_update` into a private module-level funciton
- split out each kind of update test (every test focuses on a single
  accessor's behavior) into its own test function

The resulting test suite was dramatically easier for me to debug when
I was understanding the problems caused by a refactor to do lazy loading.

Reviewed By: grievejia

Differential Revision: D35601392

fbshipit-source-id: d9097ebbd8ac1b5c0b0ff7f8bc15d8f2f2fe430c