Get rid of SharedMemory vs Memory aliasing
Summary:
We re-export the SharedMemory table interface from Memory; the idea is
that this allows us to control all access so that if we wanted to swap
out the implementation it would be feasible to do so.

As a result, we ought to rely on Memory rather than SharedMemory everywhere
in the downstream code.

This is the case, but it gets obfuscated because we alias Memory to
SharedMemory in a handful of the interprocedural modules, in effect undoing
the re-export! This can be a source of confusion, it made my life harder
when I tried to track down all uses of shared memory tables.

Let's not do that - call the Memory module Memory everywhere it is used.

Reviewed By: arthaud

Differential Revision: D35571025

fbshipit-source-id: 8465dcba72df9c6e601da30c2c95a9748caf97ea