Do not give up the analysis on untracked types
Summary:
Type checking can throw the `ClassHierarchy.Untracked` exception when encountering a type that is not defined in the environment.
In the taint analysis, we catch this exception at the top level of the call graph construction and the global fixpoint analysis, which means we will entirely skip a function if that exception is thrown at any point.

This diff changes the logic so that we catch the exception closer to where it is raised, so we can have a proper default behavior and keep analyzing the current callable.

Differential Revision: D35596829

fbshipit-source-id: a3f28da233e3fa19bcfceabc815ac3d19ba42132