Add `qualifier` and `is_in_project` to `SourcePath`
Summary:
To write the AstEnvironment with a lazy update will require
mapping or filtering source paths to get the qualifier or see
if the source is in the project in quite a few places.

There's quite a bit less line noise if we just expose these as
normal functions in SourcePath so that we don't have to make
lambdas that destructure the source path repeatedly.

Reviewed By: grievejia

Differential Revision: D35617422

fbshipit-source-id: 3b6b4cf799512d97c32ee55efa89a85c042e4c5a