Fix dual module-tracker problem in astEnvironmentTest.ml
Summary:
In the test `test_parse_sources`, we were creating two different
module trackers and passing an update produced by the second one
to an ast environment produced by the first one.

This worked by accident in the old code - the actual reads would
fail because of course the ast_environment couldn't find the
files created in new temp directories, but we produce placeholders
for failed reads and the invalidated_modules was still correct.

But the semantics were meaningless - it's never actually valid
to pass updates created from one module tracker to an environment
backed by another one. And in the lazy parsing world it produces
weird inconsistencies.

This commit is a minimal fix:
- move the definitions of the directory roots up so that the create and update
  can reuse them. Also move up the `write_file` helper function since it's the
  same in both sections.
- add `stub_root` to the initial configuration, even though it isn't used;
  adding it later is not legal, the configuration has to be immutable over
  the lifetime of a file!
- reuse the existing module tracker (and therefore also configuration) in the
  update subsection of the test
- don't use `a.py` in the update test - that name collides with the `a.pyi`
  from the create session. To keep the names clear I use `new_local.py`
  and `new_stub.pyi` for the new filenames.

The semantics here change as little as possible in order to make the test
sensible; the actual / expected results change not at all except for
filenames where necessary.

Reviewed By: grievejia

Differential Revision: D35659031

fbshipit-source-id: 3973dac2d2fd91543bd4e3f4a8b97e78f9a6b927