Skip overrides on `str`
Summary:
We mark `Markupsafe.__init__` as a `XSS` sink, and that class extends `str`.
Now that typeshed does not include stubs for `markupsafe` anymore, we analyze its code and infer (rightfully) sinks on `__add__`, `join` and other methods.
Now, because it extends `str`, whenever we call `str.__add__`, `str.join`, etc. we get sinks from `Markupsafe`.
This is legitimate in cases where the string could be a `Markupsafe` (e.g, `def foo(x: str):`) but wrong for string literals (e.g, `"|".join(x)`).
To work around this problem, let's mark `str` as `SkipOverrides`, which means we won't join models from its overrides.
This is what we want in most cases.

Reviewed By: gbleaney

Differential Revision: D35663681

fbshipit-source-id: 04a0917fae64f20adccdf49a16f862cc63a6c581