Add arguments to explicitly preserve source or sink sanitizer transforms
Summary:
When handling taint transforms, we have to be careful when we want to keep source or sink sanitizer transforms.
For instance, when adding sanitize taint transforms on a non-tito sink, we only want to keep the source sanitizers. When adding sanitize taint transforms on a tito sink, we want to keep both.

To make this more explicit in the code, let's add two parameters `~preserve_sanitize_sources` and `~preserve_sanitize_sinks`.

We also know that D35484441 (https://github.com/facebook/pyre-check/commit/4eaf497eebc544716c2663b2c8b7bc0f88712508) has been increasing memory usage, and my random guess is that creating a temporary list in `Sinks.apply_transforms` might be the reason. This diff might fix that problem.

Differential Revision: D35682139

fbshipit-source-id: c3dd0e47a2d9f37559b378e25d13c55e2bda1447