Implement a `FilterMap` transform operator
Summary:
We have `Filter : 'a -> bool` and `Map : 'a -> 'a` operators on abstract domains, but we do not have a `FilterMap : 'a -> 'a option` operator.
In theory, one could use a `Seq(Map, Filter)` but the problem with that approach is that it requires the first `Map` to return the same type, which might not have an element to represent an invalid value.

Let's implement a `FilterMap` operator on abstract domains to get around that problem. It will also be more efficient.

This will be used in a follow-up.

Reviewed By: dkgi

Differential Revision: D35694451

fbshipit-source-id: f6d3f9c47914d997a7a3e9048db1906e5a49ff97