Forward analysis of `.format()` and % string literal implicit sinks
Summary:
https://pyre-check.org/docs/pysa-advanced/#literal-string-sources-and-sinks

Formatted strings are already literal string sinks, but `"<{}>".format(s)` and `"<%s>" % (s,)` do not treat the string as an implicit sink matching the regex `^<.*>$`, while it would be the case for `f"<{s}>`.

This diff helps achieve feature parity and catch past issues that were missed for SQL string literals.

In addition, tito information is properly added for format strings, `.format`, and % formatting.

Reviewed By: arthaud

Differential Revision: D35570310

fbshipit-source-id: 7fbf2bdb429ef82457657b6b63ad229f68a3318c