lazy-parsing: Get all possible owning qualifiers for a (non-qualifier) reference
Summary:
In order to make UnannotatedGlobalEnvironment lazy, we will need the ability
to take a reference that does not yet exist in one of the lookup maps
(for example, the FunctionDefinitions shared memory table) and load
it.

Because whether the maps are expected to contain all valid keys depends on
whether the module has been loaded yet, we need to check all the possible containing
qualifiers - due to nested classes, it's not possible to naively convert a reference into
a single qualifier.

This function will let us do so - it gives all the potential qualifiers for a given name.
It doesn't include the empty qualifier (`builtins.pyi`) because builtins does not include
nested classes and likely never will.

Reviewed By: grievejia

Differential Revision: D35659548

fbshipit-source-id: c2f94ac46c81f89e35e1871a9e89f6fde24011f2