lazy-parser: Make UnannotatedGlobalEnvironment lazy in theory
Summary:
**What this commit does**

This commit causes UnannotatedGlobalEnvironment to use lazy,
read-only views of the dependency tracked tables that back it. This
can in theory make the environment lazy, although that change won't
actually take effect until we also update AstEnvironment to be lazy:
because AstEnvironment still eagerly parses everything and passes
along everything as `invalidated_modules` the laziness in this commit
is mostly hypothetical.

**How this commit works at a high level**

The thing that makes UnannotatedGlobalEnvironment very different
from other environments is that the computation granularity - a full module,
because we only want to pay the preprocessing cost once - is different
from the read granularity of individual classes, functions, etc.

We want to make it "behave like" EnvironmentTable, which uses
`produce_value` to create a lazy read-only cache of computations, but
we have to implement it differently because we need to ensure that
reads on *any* key trigger us to process the whole module.

We use `LazyLoading.load_module_if_exists` to handle the module
reads (it just wraps `set_module_data` with a check to make sure the
module is actually missing).

A further complication is that because we use fully-qualified names
that don't know which part of the name is the module portion, we
cannot actually deterministically load the right module whenever
the initial check of a lazy read returns `None`; instead, we have
to try loading every possible containing qualifier (every prefix
of the reference). This is handled by `LazyLoading.load_all_possible_modules`.

**How this commit works in detail**

We have a two-step process for both lazy `get` and lazy `mem`.

It works like this:
- First, try calling `get` or `mem`
  - if we get a clear hit `Some _` or `true`, then return
  - otherwise, use `LazyLoading.load_all_possible_modules` to
    load all the potential qualifiers that haven't been loaded
    - then, call `get` or `mem` again and return whatever the
      result is

This is basically the same as the way `EnvironmentTable.get` works
in terms of `produce_value`; the main difference is that instead of
just producing the value we need, in this case the underlying
computation function produces quite a lot of extra information
as well. So instead of trying to directly return the result we just
look it up after we've loaded everything.

**Perf concerns**

For the most part, the lazy-loaded environment should perform
similarly to the eagerly-loaded environment from before this commit.
Reads of keys that already exist should be just as fast, and reads
of new keys will process just the ancestor modules; we'll never
load the same module twice as we do this (other than races between
workers when running in parallel; in those cases the first worker
will always win).

The only operation that changes cost significantly is trying
to read a name that doesn't exist. Because we don't store nonexistent
keys, we always have to check whether some prefix reference
corresponds to a module we haven't loaded yet; this is true even
for duplicate reads of the same key. That lookup is still `O(1)`
(assuming the reference length is considered bounded by a constant)
and won't require any deserialization so it should be cheap,
but it's going to be quite a bit slower than just a single `mem`
operation.

As a result, we'll want to check on the perf results - there
are ways to mitigate the problem, but I'd rather ignore it unless
there's a significant regression.

Also it's worth noting that the perf hit (and quite a bit of the
complexity from scanning prefixes) would go away if Reference.t
knew the difference between the module qualifier and in-module name;
that's a nasty refactor but we should really do it, since lacking it
also causes nondeterminism in `pyre incremental` (see T116949717).

Differential Revision: D35659550

fbshipit-source-id: 45043059e49b1638abf8e6ce5d5a67c2b1bbd4a8