lazy-parser: Make UnannotatedGlobalEnvironment use lazy lookups
Summary:
**What this commit does**

This commit causes UnannotatedGlobalEnvironment to use lazy,
read-only views of the dependency tracked tables that back it. This
can in theory make the environment lazy, although that change won't
actually take effect until we also update AstEnvironment to be lazy:
because AstEnvironment still eagerly parses everything and passes
along everything as `invalidated_modules` the laziness in this commit
is mostly hypothetical.

In the next diff D35659549 we actually trigger the laziness by making
AstEnvironment lazy and passing a much smaller set of invalidated
modules (only in-project sources on cold start, and in-project sources
+ all previously-read sources that changed on updates).

**How this commit works at a high level**

The thing that makes UnannotatedGlobalEnvironment very different
from other environments is that the computation granularity - a full module,
because we only want to pay the preprocessing cost once - is different
from the read granularity of individual classes, functions, etc.

We want to make it "behave like" EnvironmentTable, which uses
`produce_value` to create a lazy read-only cache of computations, but
we have to implement it differently because we need to ensure that
reads on *any* key trigger us to process the whole module.

We use `LazyLoading.load_module_if_exists` to handle the module
reads (it just wraps `set_module_data` with a check to make sure the
module is actually missing).

A further complication is that because we use fully-qualified names
that don't know which part of the name is the module portion, we
cannot actually deterministically load the right module whenever
the initial check of a lazy read returns `None`; instead, we have
to try loading every possible containing qualifier (every prefix
of the reference). This is handled by `LazyLoading.load_all_possible_modules`.

**How this commit works in detail**

We have a two-step process for both lazy `get` and lazy `mem`.

It works like this:
- First, try calling `get` or `mem`
  - if we get a clear hit `Some _` or `true`, then return
  - otherwise, use `LazyLoading.load_all_possible_modules` to
    load all the potential qualifiers that haven't been loaded
    - then, call `get` or `mem` again and return whatever the
      result is

This is basically the same as the way `EnvironmentTable.get` works
in terms of `produce_value`; the main difference is that instead of
just producing the value we need, in this case the underlying
computation function produces quite a lot of extra information
as well. So instead of trying to directly return the result we just
look it up after we've loaded everything.

**Code cleanliness**

There's some copypasta introduced by this commit. Each of the
tables backing UnannotatedGlobalEnvironment is now a struct
that includes the dependency tracked table but also adds some other
information needed to do the lazy wrapping (`key_to_reference`, which
is needed because `ClassSummaries` is keyed on `Type.Primitive.t = string`
rather than reference, and `is_qualifier`, which tells the lazy lookup whether
the reference itself needs to be considered a qualifier or just its prefixes).

I can work on cleaning that up but I'd rather do it in a separate commit,
this logic is pretty tricky and I think it's better to be clear about the change
and have messier code then to make beautiful abstractions that make
it harder to see exactly how the algorithm works relative to the previous
code.

**Perf concerns**

For the most part, the lazy-loaded environment should perform
similarly to the eagerly-loaded environment from before this commit,
but there is a perf hit.

The only operation that changes cost significantly is trying
to read a name that doesn't exist. Because we don't store nonexistent
keys, we always have to check whether some prefix reference
corresponds to a module we haven't loaded yet; this is true even
for duplicate reads of the same key. That lookup is still `O(1)`
(assuming the reference length is considered bounded by a constant)
and won't require any deserialization so it should be cheap,
but it's going to be quite a bit slower than just a single `mem`
operation.

Reviewed By: grievejia

Differential Revision: D35659550

fbshipit-source-id: b1d23a7d893acef56b0d8881c22a29652a470238