lazy parsing: Make AstEnvironment lazy
Summary:
This commit makes AstEnvironment lazy.

We do this in a few steps:
- Create a new `LazyRawSources` module that is a read-only lazy view of
  `RawSources` with just a `get` function.
- Force the read-only record in `AstEnvironment.read_only` rely
  exclusively on `LazyRawSources.get` rather than `RawSources.get`.
- Change the update logic:
  - a cold start update should parse nothing, but invalidate all
    internal modules (type check relies on the invalidated modules
    list to know what to check).
  - both shallow and fine grained updates should
    - clear all modules that changed:
      - modules requiring reparse
      - new implicit modules, which might come from a deleted __init__
      - deleted modules, which need to be wiped
    - reparse all modules corresponding to changed files if those
      modules already exist in RawSources
  - updates should return as implicit modules both all modules
    that changed, plus all internal modules with changes (this
    is needed for new internal paths, otherwise we'll fail to
    run typecheck on them downstream), plus all modules whose
    preprocessing step registered a dependency on a changed
    raw source.

**Perf implications**

It looks like on a large project Pyre gets 5-15% faster. It is
not yet clear how much of this comes from parsing we don't have
to do at all versus improved ability to mix compute-bound
type checking with IO-bound parsing.

NOTE: I found a preexisting bug when writing this diff: it is not
possible to make incremental mode fully consistent in Shallow mode.
Without dependency tracking, we don't know when a processed Ast
has been invalidated by an upstream RawSource (this would happen
if there is a wildcard import). This won't matter for data
consistency because we'll wipe UnannotatedGlobalEnvironment, but it
*does* potentially matter for typecheck consistency because we only
re-check invalidated modules. So if a change to wildcard-imported
values causes typecheck results to fail, we'll miss that because
we'll never rerun the check!
...
This is another reason why we should consider explicitly getting rid of
Shallow-mode updates, either by making the environment always build
with dependency tracking or by separating the create and update logic
and throwing an exception when someone tries to update in Shallow
mode.
...
It's not an urgent bug to fix in this diff because incremental
updates in Shallow mode are a dead code path - the logic exists (and
is buggy) but there's no way to actually trigger it.

Reviewed By: grievejia

Differential Revision: D35659549

fbshipit-source-id: ee78b7d4a2f69c803ef9ac02845ab157321d951c