Backward analysis of `.format()` and % string literal implicit sinks
Summary:
See context here: D35570310 (https://github.com/facebook/pyre-check/commit/631331fae77b757c26e5f5c34688372d4ecf65a0)

## Implementation details

To reuse code I constructed a list of string parts that included the formatted string and the arguments as expressions.

To make it work, I refactored the turn f string into a regular string to check regex on code from turning the formatted expressions into `"{}"` to `""` so as to not add extra `{}` to the end of the resulting regular string.

Reviewed By: arthaud

Differential Revision: D35664034

fbshipit-source-id: b18c42fab1ecd723ccdc4c8bdbfdd9b681e223dc